separating the tests that use Daniela's topology
